Model: Add support for dynamic model selection
Add support for dynamic selection and loading of models with the `Model`
struct and for declaring new models with runtime functionality with the
`model.VTable` helper.

This API was inspired by `std.mem.Allocator`'s API which permits outside
users to define their own `Allocator` implementations. This will be
fully supported by both the supported GPT-2 and Llama model
implementations in the near future.